[IMP] bridge_seller_delivery_amazon, seller_management : Avoid amazon… (#55)
* [IMP] bridge_seller_delivery_amazon, seller_management : Avoid amazon document pdf generation if there is no amazon shipping orders. Remove additional fields in DO for self fulfilment seller

* [FIX] bridge_seller_delivery_amazon : removed unwanted fields

* [FIX] seller_management : removed unwanted fields

* Update stock_picking.py

* Update stock_picking.py

* Update stock_picking.py

+++ bridge_seller_delivery_amazon/models/bulk_process.py
@@ -29,29 +29,33 @@ class BulkProcessInherit(models.Model):
     def generate_amzn_shipment_document(self):
         self._check_amzn_stock_picking_ids()
         merged_pdf = PdfFileMerger()
-        for picking in self.stock_picking_ids.filtered(lambda r: r.carrier_id.name == 'amazon'):
-            picking.amzn_get_shipment_document()
-            # Attach the PDF to the bulk process attachments
-            pdf_attachment = self._attach_pdf_from_picking(picking)
-            # Merge the PDF into the merged_pdf object
-            pdf_content = base64.b64decode(pdf_attachment.datas)
-            merged_pdf.append(fileobj=BytesIO(pdf_content))
-            pdf_attachment.unlink()  # Remove the temporary PDF attachment
-        # Write the merged PDF to a new attachment
-        merged_pdf_bytes = BytesIO()
-        merged_pdf.write(merged_pdf_bytes)
-        merged_pdf_bytes.seek(0)
-        merged_pdf_content = base64.b64encode(merged_pdf_bytes.getvalue())
-        merged_pdf_bytes.close()
-        self.env['ir.attachment'].create({
-            'name': 'Amazon Label Document',
-            'datas': merged_pdf_content,
-            'res_model': "shiprocket.bulk.process",
-            'res_id': self.id,
-        })
-        merged_pdf.close()  # Close the merged PDF object
-        self.is_document_generated = True
-        self.state = 'Document Generated'
+        pickings = self.stock_picking_ids.filtered(lambda r: r.carrier_id.name == 'amazon')
+        if not pickings:
+            raise UserError(('No Amazon Pickings To Proceed'))
+        else:
+            for picking in pickings:
+                picking.amzn_get_shipment_document()
+                # Attach the PDF to the bulk process attachments
+                pdf_attachment = self._attach_pdf_from_picking(picking)
+                # Merge the PDF into the merged_pdf object
+                pdf_content = base64.b64decode(pdf_attachment.datas)
+                merged_pdf.append(fileobj=BytesIO(pdf_content))
+                pdf_attachment.unlink()  # Remove the temporary PDF attachment
+            # Write the merged PDF to a new attachment
+            merged_pdf_bytes = BytesIO()
+            merged_pdf.write(merged_pdf_bytes)
+            merged_pdf_bytes.seek(0)
+            merged_pdf_content = base64.b64encode(merged_pdf_bytes.getvalue())
+            merged_pdf_bytes.close()
+            self.env['ir.attachment'].create({
+                'name': 'Amazon Label Document',
+                'datas': merged_pdf_content,
+                'res_model': "shiprocket.bulk.process",
+                'res_id': self.id,
+            })
+            merged_pdf.close()  # Close the merged PDF object
+            self.is_document_generated = True
+            self.state = 'Document Generated'
 
     def _attach_pdf_from_picking(self, picking):
         if picking.barcode:
